rename dataset and trainer
