+++ tests/test_integration.py
@@ -6,7 +6,7 @@ import h5py
 from deeprankcore.preprocess import preprocess
 from tests._utils import PATH_TEST
 from deeprankcore.query import ProteinProteinInterfaceResidueQuery
-from deeprankcore.dataset_ import GraphDataset
+from deeprankcore.dataset import GraphDataset
 from deeprankcore.trainer import Trainer
 from deeprankcore.neuralnets.ginet import GINet
 from deeprankcore.utils.metrics import OutputExporter
